Add tables of correspondences to orient users during migrations
This is a first step toward the implementation of the
tables of correspondences.

These tables are compass to orient users during their migrations.
Their goal is to help people to identify alternatives case by case.

These tables will be completed later with links toward the
alternatives docs/repos.

These tables will be completed later with some succint examples
of alternatives.