wsgi: No request Content-Length nor Transfer-Encoding implies no body
Previously, if an application tried to read some set length from
such a request (because, say, it always tried to read every
request's body), the server could hang waiting on the client to
send more bytes.

The spec, however, is clear -- even going back to HTTP/1.0 -- if
Content-Length (nor Transfer-Encoding, in HTTP/1.1) is missing,
the request has no body, which is equivalent to having a body of
length 0. See:

- https://www.rfc-editor.org/rfc/rfc1945#section-7.2
- https://www.rfc-editor.org/rfc/rfc2616#section-4.3
- https://www.rfc-editor.org/rfc/rfc7230#section-3.3
- https://www.rfc-editor.org/rfc/rfc9112#section-6.3-2.7