Spew: Correct line lookup from inspect.getsourcelines()
I can't write a test for this, as I don't know how to trigger is part of
Spew, without also falling into the OSError branch (which
test_line_nofile tests).

However, it has been observed to hit this code when running the tests
under pytest-xdist (execnext), and the code is obviously wrong. This
corrects it.