Discontinues PyPy support
PyPy broke our CI and avoid us merging important patches.

We are not PyPy users, and we have no enough resources to spend
days maintaining compatibility with PyPy. We want to focus our
time on our use cases.

Persons interested to continue maintaining the PyPy support are
welcomed, and we could surely consider reintroducing its support
if people volunteer to maintain it.

For now our priority is to ensure that we are able to offer a
stable migration to our users, so we first have to ensure that
we remain safe.